better user-search

+++ src/components/repositories/Repositories.js
@@ -1,13 +1,18 @@
 import React, { useEffect, useState } from 'react'
 import * as S from './styled'
+import { useHistory } from 'react-router-dom'
 
 export default function Repositories() {
+    const history = useHistory()
     const [repositories, setRepositories] = useState([])
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName')
-        repositoriesName = JSON.parse(repositoriesName)
-        setRepositories(repositoriesName)
-        localStorage.clear()
+        if (repositoriesName !== null) {
+            repositoriesName = JSON.parse(repositoriesName)
+            setRepositories(repositoriesName)
+            localStorage.clear()
+        }
+        else {history.push('/user-search')}
     }, [])
 
     return (
@@ -19,6 +24,7 @@ export default function Repositories() {
                         <S.ListItem>{repository}</S.ListItem>
                     )})}
             </S.List>
+            <S.LinkHome to='/user-search'>Voltar</S.LinkHome>
         </S.Container>
     )
 }
